Final Update Minus Fixes (Some auth related ones preventing me from viewing parts of navigation)

+++ app/Http/Controllers/JobController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Mail\JobPosted;
 use App\Models\Job;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 
@@ -34,19 +36,31 @@ class JobController extends Controller
     
         ]);
     
-        Job::create([
+        $job = Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
+
+        Mail::to($job->employer->user)->queue(
+            new JobPosted($job)
+        );
     
         return redirect('/jobs');
     }
 
-    public function edit(Job $job){
+    // public function edit(Job $job){
 
-            Gate::authorize('edit-job', $job);
+    //         Gate::authorize('edit-job', $job);
 
+
+    //     return view('jobs.edit', ['job' => $job]);
+    // }
+
+
+    public function edit (Job $job){
+
+        Gate::authorize('edit-job', $job);
 
         return view('jobs.edit', ['job' => $job]);
     }
@@ -72,7 +86,10 @@ class JobController extends Controller
         return redirect('/jobs/'.$job->id);
     }
 
-    public function delete(Job $job){
+    public function destroy(Job $job){
+
+        Gate::authorize('edit-job', $job);
+
         $job->delete();
 
         return redirect('/jobs');
